Some refactoring an enhancements for mobile

+++ Common/GetAllLocationsForUser_Result.cs
@@ -21,9 +21,9 @@ namespace Common
         public string State { get; set; }
         public Nullable<int> Zipcode { get; set; }
         public string LattitudeDirection { get; set; }
-        public Nullable<decimal> Lattitude { get; set; }
+        public Nullable<double> Lattitude { get; set; }
         public string LongitudeDirection { get; set; }
-        public Nullable<decimal> Longitude { get; set; }
+        public Nullable<double> Longitude { get; set; }
         public System.DateTime CreateDate { get; set; }
     }
 }
